src/python/requirements.txt
zensols.actioncli>=1.0.6
gitpython>=2.1.11
